Extract checkout steps machinery to EdenflowersWeb.CheckoutComponents
The steps/1 + checkout_step/1 + step_title + step_label + step_summary +
format_date set is ~125 lines describing how the checkout's step list
renders. It's self-contained — takes a state atom and an order, emits
the edit_step event via JS.push, and uses inner_block for the active
step's form.

Moving it out leaves the LV focused on state-machine transitions, form
submission, Stripe orchestration, and the card drawer. The component
module owns step rendering exclusively, and the call site stays
unchanged thanks to `import ..., only: [steps: 1]`.
